deprecate custom HtmlFormatter for jsondiffpatch  - used to hot-fix DOM-XSS bug
